server/requirements/prod.txt
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
python-dateutil==2.8.0
gunicorn==19.9.0
pylint==2.3.1
psycopg2-binary==2.8.3
